# frontend/contracts/openapi/backend_openapi.json
{
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "properties": {
          "access_token": {
            "minLength": 1,
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": ["access_token"],
        "title": "AccessTokenResponse",
        "type": "object"
      },
...
      "AuthenticatedUserResponse": {
        "properties": {
          "disabled": {
            "title": "Disabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "username": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": ["id", "username", "disabled"],
        "title": "AuthenticatedUserResponse",
        "type": "object"
      },
      "Body_login_for_access_token_v1_token_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "title": "Body_login_for_access_token_v1_token_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterUserRequest": {
        "properties": {
          "password": {
            "maxLength": 255,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "title": "RegisterUserRequest",
        "type": "object"
      },
...
      "UpdateCurrentUserRequest": {
        "properties": {
          "current_password": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Password"
          },
          "new_password": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Password"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "title": "UpdateCurrentUserRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/token": {
      "post": {
        "description": "Authenticate a user with `application/x-www-form-urlencoded` and return a JWT bearer token.",
        "operationId": "login_for_access_token_v1_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "example": {
                "password": "{{password}}",
                "username": "demo_user"
              },
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_v1_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "token_type": "bearer"
                },
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            },
            "description": "Authentication successful."
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "invalid_input",
                  "detail": "Request validation error",
                  "meta": [
                    {
                      "loc": ["body", "username"],
                      "msg": "Field required"
                    }
                  ],
                  "status": 400
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Invalid credential input format."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "unauthorized",
                  "detail": "Invalid username or password",
                  "status": 401
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Invalid username or password.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Inactive user",
                  "status": 403
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "User is authenticated but inactive."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "internal_error",
                  "detail": "Internal server error",
                  "status": 500
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Unhandled internal server error."
          }
        },
        "summary": "Get access token",
        "tags": ["auth"]
      }
    },
    "/v1/users/me": {
      "get": {
        "description": "Return the user associated with the bearer token in the `Authorization` header.",
        "operationId": "read_current_user_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "disabled": false,
                  "id": 1,
                  "permissions": [
                    "role_permissions:manage",
                    "roles:manage",
                    "user_roles:manage",
                    "users:manage"
                  ],
                  "username": "admin"
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            },
            "description": "Authenticated user."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "unauthorized",
                  "detail": "Could not validate credentials",
                  "status": 401
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Missing, expired, or invalid token.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Inactive user",
                  "status": 403
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "User is authenticated but inactive."
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "internal_error",
                  "detail": "Internal server error",
                  "status": 500
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Unhandled internal server error."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get current user",
        "tags": ["auth"]
      },
      "patch": {
        "description": "Update `username`, password, or both. Password change requires both `current_password` and `new_password`.",
        "operationId": "update_current_user_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrentUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "disabled": false,
                  "id": 1,
                  "permissions": [],
                  "username": "profile_user_v2"
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            },
            "description": "Profile updated."
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "invalid_input",
                  "detail": "At least one field must be provided to update the user",
                  "status": 400
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Invalid request or no effective changes."
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "unauthorized",
                  "detail": "Current password is invalid",
                  "status": 401
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Invalid token or incorrect current password.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Inactive user",
                  "status": 403
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "User is authenticated but inactive."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": "conflict",
                  "detail": "Username already exists",
                  "meta": {
                    "username": "admin"
                  },
                  "status": 409
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "New username is already in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "internal_error",
                  "detail": "Internal server error",
                  "status": 500
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Unhandled internal server error."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update current user",
        "tags": ["auth"]
      }
    },
    "/v1/users/register": {
      "post": {
        "description": "Create a local active user. `username` is normalized to lowercase and the password must satisfy the service password policy.",
        "operationId": "register_user_v1_users_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "disabled": false,
                  "id": 1,
                  "permissions": [
                    "role_permissions:manage",
                    "roles:manage",
                    "user_roles:manage",
                    "users:manage"
                  ],
                  "username": "admin"
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            },
            "description": "User registered."
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "invalid_input",
                  "detail": "Password does not meet policy",
                  "meta": {
                    "violations": [
                      "Password must include at least one uppercase letter"
                    ]
                  },
                  "status": 400
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Invalid registration input or password policy violation."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": "conflict",
                  "detail": "Username already exists",
                  "meta": {
                    "username": "admin"
                  },
                  "status": 409
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Username already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "internal_error",
                  "detail": "Internal server error",
                  "status": 500
                },
                "schema": {
                  "properties": {
                    "code": {
                      "description": "Machine-readable domain error code.",
                      "type": "string"
                    },
                    "detail": {
                      "description": "Human-readable error message.",
                      "type": "string"
                    },
                    "meta": {
                      "description": "Optional structured error metadata."
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "type": "integer"
                    }
                  },
                  "required": ["detail", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Unhandled internal server error."
          }
        },
        "summary": "Register user",
        "tags": ["auth"]
      }
    }
  }
}